Changes and features added:
A slider was added into the A2DP screen of the app, giving somewhat complete control over the system volume. See .kv file for limitations...
Created a function that allows for python to alter system volume, which in turn, is controlled by a slider in the A2DP screen.
Added a new file: volumelevel. This stores the system volume level in a file for reading later.